refactor: sophiah.py to sophia.py
